requirements.txt
# GUI 框架 - 使用兼容版本的 PySide6
PySide6==6.5.3

# 计算机视觉和图像处理
opencv-python==4.8.1.78
numpy>=1.21.0

# 网络和通信
requests>=2.25.0
pyserial==3.5

# 手势识别
mediapipe==0.10.7

# 系统工具
psutil>=5.8.0

# 安全和加密
keyring>=23.0.0
cryptography>=3.4.0
packaging>=21.0

# 火山引擎语音识别依赖
websockets==11.0.3
pyaudio>=0.2.11
